+++ CA7/DeviceTree/WildhornAV/u-boot/stm32mp15-mx.dtsi
@@ -14,7 +14,7 @@
  * Relaxed Timing Mode: false
  * Address mapping type: RBC
  *
- * Save Date: 2022.02.20, save Time: 18:21:51
+ * Save Date: 2022.02.20, save Time: 18:45:45
  */
 
 #define DDR_MEM_NAME	"DDR3-DDR3L 16bits 533000kHz"
