+++ CM4/Core/Src/ipcc.c
@@ -1,0 +1,97 @@
+/**
+  ******************************************************************************
+  * @file    ipcc.c
+  * @brief   This file provides code for the configuration
+  *          of the IPCC instances.
+  ******************************************************************************
+  * @attention
+  *
+  * <h2><center>&copy; Copyright (c) 2022 STMicroelectronics.
+  * All rights reserved.</center></h2>
+  *
+  * This software component is licensed by ST under Ultimate Liberty license
+  * SLA0044, the "License"; You may not use this file except in compliance with
+  * the License. You may obtain a copy of the License at:
+  *                             www.st.com/SLA0044
+  *
+  ******************************************************************************
+  */
+
+/* Includes ------------------------------------------------------------------*/
+#include "ipcc.h"
+
+/* USER CODE BEGIN 0 */
+
+/* USER CODE END 0 */
+
+IPCC_HandleTypeDef hipcc;
+
+/* IPCC init function */
+void MX_IPCC_Init(void)
+{
+
+  /* USER CODE BEGIN IPCC_Init 0 */
+
+  /* USER CODE END IPCC_Init 0 */
+
+  /* USER CODE BEGIN IPCC_Init 1 */
+
+  /* USER CODE END IPCC_Init 1 */
+  hipcc.Instance = IPCC;
+  if (HAL_IPCC_Init(&hipcc) != HAL_OK)
+  {
+    Error_Handler();
+  }
+  /* USER CODE BEGIN IPCC_Init 2 */
+
+  /* USER CODE END IPCC_Init 2 */
+
+}
+
+void HAL_IPCC_MspInit(IPCC_HandleTypeDef* ipccHandle)
+{
+
+  if(ipccHandle->Instance==IPCC)
+  {
+  /* USER CODE BEGIN IPCC_MspInit 0 */
+
+  /* USER CODE END IPCC_MspInit 0 */
+    /* IPCC clock enable */
+    __HAL_RCC_IPCC_CLK_ENABLE();
+
+    /* IPCC interrupt Init */
+    HAL_NVIC_SetPriority(IPCC_RX1_IRQn, 5, 0);
+    HAL_NVIC_EnableIRQ(IPCC_RX1_IRQn);
+    HAL_NVIC_SetPriority(IPCC_TX1_IRQn, 5, 0);
+    HAL_NVIC_EnableIRQ(IPCC_TX1_IRQn);
+  /* USER CODE BEGIN IPCC_MspInit 1 */
+
+  /* USER CODE END IPCC_MspInit 1 */
+  }
+}
+
+void HAL_IPCC_MspDeInit(IPCC_HandleTypeDef* ipccHandle)
+{
+
+  if(ipccHandle->Instance==IPCC)
+  {
+  /* USER CODE BEGIN IPCC_MspDeInit 0 */
+
+  /* USER CODE END IPCC_MspDeInit 0 */
+    /* Peripheral clock disable */
+    __HAL_RCC_IPCC_CLK_DISABLE();
+
+    /* IPCC interrupt Deinit */
+    HAL_NVIC_DisableIRQ(IPCC_RX1_IRQn);
+    HAL_NVIC_DisableIRQ(IPCC_TX1_IRQn);
+  /* USER CODE BEGIN IPCC_MspDeInit 1 */
+
+  /* USER CODE END IPCC_MspDeInit 1 */
+  }
+}
+
+/* USER CODE BEGIN 1 */
+
+/* USER CODE END 1 */
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
