+++ CM4/ert/inc/control.h
@@ -1,0 +1,57 @@
+/*  Title       : Control
+ *  Filename    : control.h
+ *  Author      : iacopo sprenger
+ *  Date        : 20.01.2022
+ *  Version     : 0.1
+ *  Description : control
+ */
+
+#ifndef CONTROL_H
+#define CONTROL_H
+
+
+
+/**********************
+ *  INCLUDES
+ **********************/
+
+#include <stdint.h>
+
+/**********************
+ *  CONSTANTS
+ **********************/
+
+
+/**********************
+ *  MACROS
+ **********************/
+
+
+/**********************
+ *  TYPEDEFS
+ **********************/
+
+
+/**********************
+ *  VARIABLES
+ **********************/
+
+
+/**********************
+ *  PROTOTYPES
+ **********************/
+
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+
+void control_thread(void * arg);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif /* __cplusplus */
+
+#endif /* TEMPLATE_H */
+
+/* END */
