+++ CM4/ert/inc/template.h
@@ -1,0 +1,55 @@
+/*  Title       : Template
+ *  Filename    : template.h
+ *  Author      : iacopo sprenger
+ *  Date        : 20.01.2021
+ *  Version     : 0.1
+ *  Description : template for code files
+ */
+
+#ifndef TEMPLATE_H
+#define TEMPLATE_H
+
+
+
+/**********************
+ *  INCLUDES
+ **********************/
+
+#include <stdint.h>
+
+/**********************
+ *  CONSTANTS
+ **********************/
+
+
+/**********************
+ *  MACROS
+ **********************/
+
+
+/**********************
+ *  TYPEDEFS
+ **********************/
+
+
+/**********************
+ *  VARIABLES
+ **********************/
+
+
+/**********************
+ *  PROTOTYPES
+ **********************/
+
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+
+#ifdef __cplusplus
+} // extern "C"
+#endif /* __cplusplus */
+
+#endif /* TEMPLATE_H */
+
+/* END */
