+++ CM4/ert/inc/threads.h
@@ -1,0 +1,57 @@
+/*  Title       : Threads
+ *  Filename    : threads.h
+ *  Author      : iacopo sprenger
+ *  Date        : 20.02.2022
+ *  Version     : 0.1
+ *  Description : threads startup code
+ */
+
+#ifndef THREADS_H
+#define THREADS_H
+
+
+
+/**********************
+ *  INCLUDES
+ **********************/
+
+#include <stdint.h>
+
+/**********************
+ *  CONSTANTS
+ **********************/
+
+
+/**********************
+ *  MACROS
+ **********************/
+
+
+/**********************
+ *  TYPEDEFS
+ **********************/
+
+
+/**********************
+ *  VARIABLES
+ **********************/
+
+
+/**********************
+ *  PROTOTYPES
+ **********************/
+
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+
+void threads_init(void);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif /* __cplusplus */
+
+#endif /* THREADS_H */
+
+/* END */
