+++ CM4/ert/src/template.c
@@ -1,0 +1,44 @@
+/*  Title		: Template
+ *  Filename	: template.c
+ *	Author		: iacopo sprenger
+ *	Date		: 20.01.2021
+ *	Version		: 0.1
+ *	Description	: template for code files
+ */
+
+/**********************
+ *	INCLUDES
+ **********************/
+
+#include "template.h"
+
+/**********************
+ *	CONSTANTS
+ **********************/
+
+
+/**********************
+ *	MACROS
+ **********************/
+
+
+/**********************
+ *	TYPEDEFS
+ **********************/
+
+
+/**********************
+ *	VARIABLES
+ **********************/
+
+
+/**********************
+ *	PROTOTYPES
+ **********************/
+
+
+/**********************
+ *	DECLARATIONS
+ **********************/
+
+/* END */
