--- CM4/ert/threads.c
@@ -1,6 +1,0 @@
-/*
- * EPFL Rocket team
- * created on: 29.11.2021
- * created by: AV - software
- *
- * */
