+++ WildhornAV.ioc
@@ -666,17 +666,17 @@ PG9.Signal=USART6_RX
 PinOutPanel.CurrentBGAView=Top
 PinOutPanel.RotationAngle=0
 ProjectManager.AskForMigrate=true
-ProjectManager.BackupPrevious=false
+ProjectManager.BackupPrevious=true
 ProjectManager.CompilerOptimize=6
 ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=false
+ProjectManager.CoupleFile=true
 ProjectManager.CustomerFirmwarePackage=
 ProjectManager.DefaultFWLocation=true
 ProjectManager.DeletePrevious=true
 ProjectManager.DeviceId=STM32MP157DABx
 ProjectManager.DeviceTreeLocation=
 ProjectManager.FirmwarePackage=STM32Cube FW_MP1 V1.4.0
-ProjectManager.FreePins=false
+ProjectManager.FreePins=true
 ProjectManager.HalAssertFull=false
 ProjectManager.HeapSize=0x200
 ProjectManager.KeepUserCode=true
@@ -691,9 +691,11 @@ ProjectManager.ProjectName=WildhornAV
 ProjectManager.RegisterCallBack=
 ProjectManager.StackSize=0x400
 ProjectManager.TargetToolchain=STM32CubeIDE
+ProjectManager.TemplateDestinationPath=/home/iacopo/Documents/ERT/Software/WildhornAV
+ProjectManager.TemplateSourcePath=/opt/st/stm32cubeide_1.7.0/plugins/com.st.stm32cube.common.mx_6.3.0.202107141111/db/extra_templates
 ProjectManager.ToolChainLocation=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_ETZPC_Init-ETZPC-false-HAL-true,3-MX_GPIO_Init-GPIO-false-HAL-true,4-MX_ADC1_Init-ADC1-false-HAL-true,5-MX_IPCC_Init-IPCC-false-HAL-true,6-MX_DAC1_Init-DAC1-false-HAL-true,7-MX_FDCAN1_Init-FDCAN1-false-HAL-true,8-MX_FDCAN2_Init-FDCAN2-false-HAL-true,9-MX_I2C1_Init-I2C1-false-HAL-true,10-MX_I2C2_Init-I2C2-false-HAL-true,11-MX_I2C5_Init-I2C5-false-HAL-true,12-MX_TIM4_Init-TIM4-false-HAL-true,13-MX_TIM5_Init-TIM5-false-HAL-true,14-MX_USART2_UART_Init-USART2-false-HAL-true,15-MX_USART3_UART_Init-USART3-false-HAL-true,16-MX_USART6_UART_Init-USART6-false-HAL-true,0-MX_PWR_Init-PWR-false-HAL-true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_ETZPC_Init-ETZPC-false-HAL-true,3-MX_GPIO_Init-GPIO-false-HAL-true,4-MX_ADC1_Init-ADC1-false-HAL-true,5-MX_IPCC_Init-IPCC-false-HAL-true,6-MX_DAC1_Init-DAC1-false-HAL-true,7-MX_FDCAN1_Init-FDCAN1-false-HAL-true,8-MX_FDCAN2_Init-FDCAN2-false-HAL-true,9-MX_I2C1_Init-I2C1-false-HAL-true,10-MX_I2C2_Init-I2C2-false-HAL-true,11-MX_I2C5_Init-I2C5-false-HAL-true,12-MX_TIM4_Init-TIM4-false-HAL-true,13-MX_TIM5_Init-TIM5-false-HAL-true,14-MX_USART2_UART_Init-USART2-false-HAL-true,15-MX_USART3_UART_Init-USART3-false-HAL-true,16-MX_USART6_UART_Init-USART6-false-HAL-true,17-MX_TIM3_Init-TIM3-false-HAL-true,0-MX_PWR_Init-PWR-false-HAL-true
 RCC.ADCCLockSelection=RCC_ADCCLKSOURCE_PLL3
 RCC.ADCFreq_Value=100000000
 RCC.AHB1234Freq_Value=200000000
